+++ lib/active_storage_validations/analyzer/video_analyzer.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "open3"
 require_relative "shared/asv_ff_probable"
 
 module ActiveStorageValidations